Refactoring separated classes into separate files
separated each class in resourceListCreator into own file.
This not only makes it more fine granular but also
dependencies become clear.
ResourcesList <-> ResourceInfo: dependency is now one way
